requirements.txt
fastapi==0.74.1
jose==1.0.0
passlib==1.7.4
pydantic==1.9.0
python-dotenv==0.19.2
python_jose==3.3.0
SQLAlchemy==1.4.25
